+++ NEWS.md
@@ -1,6 +1,7 @@
 MixedModels v3.9.0 Release Notes
 ========================
 * Add support for `StatsModels.formula` [#536]
+* Internal method `allequal` renamed to `isconstant` [#537]
 
 MixedModels v3.8.0 Release Notes
 ========================
@@ -209,3 +210,4 @@ Package dependencies
 [#523]: https://github.com/JuliaStats/MixedModels.jl/issues/523
 [#524]: https://github.com/JuliaStats/MixedModels.jl/issues/524
 [#536]: https://github.com/JuliaStats/MixedModels.jl/issues/536
+[#537]: https://github.com/JuliaStats/MixedModels.jl/issues/537